Refactor General form build tab
